Added AnalyzerTest.RunAsync overload accepting array of expected Diagnostic results.
This will make it easier for developers having an hard time with making a function that's goal is to be used by all of their tests (even if some of them expects to test that they generate a Diagnostic while other tests might verify that they do not.

Fixes https://github.com/dotnet/roslyn-sdk/issues/860.