Provide an extensibility point in AnalyzerTest for sorting diagnostics.
The sort here is non-deterministic for any analyzers whose diagnostics can differ only on the message format and properties.

Allow the sorting algorithm to be changed by the user to include either or both of these elements if they emit diagnostics that differ only on these values from a concurrent analyzer (the .NET Interop team owns an analyzer like this, which prompted this work). This will ensure that everyone can ensure that their diagnostic and code-fix tests (which rely on the ordering to determine the order for the iterative code-fix tests and the fix-all tests) are deterministic.